refactor(epicenter): co-locate persistence with docs, remove providers folder
Move persistence code into $lib/docs/ since each persistence provider
is tightly coupled to a single doc type (registry, head, workspace).

- Move tauri-persistence.ts → docs/persistence/
- Move tauri-workspace-persistence.ts → docs/persistence/
- Rename definition-persistence.ts → docs/read-definition.ts (clearer name)
- Delete empty $lib/providers/ folder
- Update all imports to use new paths
